Move src/http1 into src/http/1.
