+++ src/client/plot-raw.js
@@ -4,7 +4,7 @@
 
 import * as Global from "../misc/Global";
 import Transaction from "./Transaction";
-import RequestParser from "../http1/RequestParser";
+import RequestParser from "../http/1/RequestParser";
 import { Must } from "../misc/Gadgets";
